+++ NEWS.md
@@ -4,6 +4,7 @@
 * `plot_all()` uses correct level-2 %NA in title
 * `simWide`: case with no observed bmi values removed
 * `traceplot()`, `densplot()`: `ncol` and `nrow` now work with `use_ggplot = TRUE`
+* `densplot()`: use of color fixed
 * functions with argument `subset` now return random effects covariance matrix correctly
 * `summary()` displayes output with rowname when only one node is returned
 * `GR_crit()`: Literature reference corrected